Changes for docker support
