--- .env.template
@@ -1,31 +1,0 @@
-# .env.template
-
-# Core Application Settings
-APPLICATIONINSIGHTS_CONNECTION_STRING="your-application-insights-connection-string"
-AZURE_AOAI_RG="your-aoai-rg-name"
-AZURE_EMBEDDINGS_VECTOR_SIZE="3072"
-AZURE_FORMREC_SERVICE="ai0-randomvalue"
-AZURE_KEY_VAULT_ENDPOINT="https://kv0-randomvalue.vault.azure.net/"
-AZURE_KEY_VAULT_NAME="kv0-randomvalue"
-AZURE_OPENAI_API_VERSION="2024-10-21"
-AZURE_OPENAI_CHATGPT_DEPLOYMENT="chat"
-AZURE_OPENAI_EMBEDDING_DEPLOYMENT="text-embedding"
-AZURE_OPENAI_EMBEDDING_MODEL="text-embedding-3-large"
-AZURE_OPENAI_SERVICE_NAME="your-aoai-service-name"
-AZURE_SEARCH_APPROACH="hybrid"
-AZURE_SEARCH_INDEX_NAME="ragindex"
-AZURE_SEARCH_SERVICE="your-ai-search-service-name"
-AZURE_STORAGE_ACCOUNT_RG="rg-0319251131"
-DOCINT_API_VERSION="2024-11-30"
-MIN_CHUNK_SIZE="100"
-NUM_TOKENS="2048"
-TOKEN_OVERLAP="200"
-MULTIMODAL="true"
-NETWORK_ISOLATION="false"
-SEARCH_ANALYZER_NAME="standard"
-SEARCH_API_VERSION="2024-07-01"
-SEARCH_INDEX_INTERVAL="PT1H"
-SEARCH_INDEX_NAME="ragindex"
-STORAGE_ACCOUNT_NAME="strag0randomvalue"
-STORAGE_CONTAINER="documents"
-STORAGE_CONTAINER_IMAGES="documents-images"
