--- app_settings.json
@@ -1,45 +1,0 @@
-{
-    "IsEncrypted": false,
-    "Values": {
-      "APPLICATIONINSIGHTS_CONNECTION_STRING": "your-appinsights-connection-string",
-      "AZURE_AOAI_RG": "rg-name-of-aoai-resource",
-      "AZURE_EMBEDDINGS_VECTOR_SIZE": "3072",
-      "AZURE_FORMREC_SERVICE": "ai0-randomvalue",
-      "AZURE_KEY_VAULT_ENDPOINT": "https://kv0-randomvalue.vault.azure.net/",
-      "AZURE_KEY_VAULT_NAME": "kv0-randomvalue",
-      "AZURE_OPENAI_API_VERSION": "2024-10-21",
-      "AZURE_OPENAI_CHATGPT_DEPLOYMENT": "chat",
-      "AZURE_OPENAI_EMBEDDING_DEPLOYMENT": "text-embedding",
-      "AZURE_OPENAI_EMBEDDING_MODEL": "text-embedding-3-large",
-      "AZURE_OPENAI_SERVICE_NAME": "azure-oai-resource-name",
-      "AZURE_SEARCH_APPROACH": "hybrid",
-      "AZURE_SEARCH_INDEX_NAME": "ragindex",
-      "AZURE_SEARCH_SERVICE": "ai-search-service-name",
-      "AZURE_STORAGE_ACCOUNT_RG": "rg-environement-name",
-      "DOCINT_API_VERSION": "2024-11-30",
-      "MIN_CHUNK_SIZE": "100",
-      "NUM_TOKENS": "2048",
-      "TOKEN_OVERLAP": "200",
-      "MULTIMODAL": "true",
-      "NETWORK_ISOLATION": "false",
-      "SEARCH_ANALYZER_NAME": "standard",
-      "SEARCH_API_VERSION": "2024-07-01",
-      "SEARCH_INDEX_INTERVAL": "PT1H",
-      "SEARCH_INDEX_NAME": "ragindex",
-      "STORAGE_ACCOUNT_NAME": "strag0randomvalue",
-      "STORAGE_CONTAINER": "documents",
-      "STORAGE_CONTAINER_IMAGES": "documents-images",
-      "AzureWebJobsFeatureFlags": "EnableWorkerIndexing",
-      "AzureWebJobsStorage__accountName": "strag0randomvalueing",
-      "AzureWebJobsStorage__credential": "managedidentity",
-      "BUILD_FLAGS": "UseExpressBuild",
-      "ENABLE_ORYX_BUILD": "true",
-      "FUNCTION_APP_NAME": "fninges0-randomvalue",
-      "FUNCTIONS_EXTENSION_VERSION": "~4",
-      "FUNCTIONS_WORKER_RUNTIME": "python",
-      "LOGLEVEL": "INFO",
-      "SCM_DO_BUILD_DURING_DEPLOYMENT": "1",
-      "XDG_CACHE_HOME": "/tmp/.cache"
-    }
-  }
-
