+++ src/connectors/images_deleted_files_purger.py
