+++ src/connectors/sharepoint/sharepoint_deleted_files_purger.py
