+++ src/connectors/sharepoint/sharepoint_files_indexer.py
