+++ src/local.settings.json.template
