+++ src/sharepoint_index_files/function.json
@@ -1,0 +1,14 @@
+{
+  "scriptFile": "__init__.py",
+  "bindings": [
+    {
+      "authLevel": "function",
+      "type": "timerTrigger",
+      "direction": "in",
+      "name": "timer",
+      "schedule": "0 */30 * * * *",
+      "runOnStartup": true,
+      "useMonitor": true
+    }
+  ]
+}
